+++ packages/web/src/components/user-list/UserList.tsx
@@ -11,8 +11,8 @@ import { getUserId } from 'common/store/account/selectors'
 import { getUsers } from 'common/store/cache/users/selectors'
 import { setNotificationSubscription } from 'common/store/pages/profile/actions'
 import * as socialActions from 'common/store/social/users/actions'
-import { makeGetOptimisticUserIdsIfNeeded } from 'common/store/tipping/selectors'
 import { loadMore, reset } from 'common/store/user-list/actions'
+import { makeGetOptimisticUserIdsIfNeeded } from 'common/store/user-list/selectors'
 import { UserListStoreState } from 'common/store/user-list/types'
 import * as unfollowConfirmationActions from 'components/unfollow-confirmation-modal/store/actions'
 import { AppState } from 'store/types'
